third_party/WebKit/LayoutTests/fast/css-generated-content/pseudo-element-events-expected.txt

PASS successfullyParsed is true

TEST COMPLETE

PASS event.target.tagName is "DIV"
PASS successfullyParsed is true

TEST COMPLETE


Pseudo elements should not dispatch events.
